requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-chroma
langchain-huggingface
langchain-groq
faiss-cpu
gtts
python-dotenv
pypdf
sentence_transformers
TTS